# backend/models/api.yaml
openapi: 3.0.3
info:
  title: Placement Portal API
  description: REST API for the Placement Portal System
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
paths:
...
  /admin/company/{company_id}:
    get:
      summary: Get company profile
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: integer
...